Revert "Changed Plugin Assembly probing from normal load to ReflectionOnlyLoad, which should have much less overhead." This reverts commit cccea98ced7e29b61998df582e771009e8063edb.
I did not think it through, and the way I wrote it, it does no longer do its job: Attributes found need to be instantiated, and also the classes these attributes are found, so that we can invoke methods there and so on. Thus we would need both, ReflectionOnly for probing if there is something potentially interesting in it. And then Load these Assemblies normally and to the same thing again on the normally loaded assembly.