[PixImage] Improve support for dual-channel images
This commit renames Col.Format.NormalUV to Col.Format.RG and
defines it to contain a red and green channel. Previously,
it was defined as having a U and V channel, whereas V is already
used for value in HSV. This change makes it possible to convert
between RG and RGB formats.